requirements.txt
# Core Python dependencies for bff-tools
PyYAML
xlsx2csv
